Support for board/categories postable_by
Adds support for the postable_by flag on boards and categories. All of
the write based post auth methods were updated to check the user's
priority against the board's, parent board's, and parent category's
postable_by to see if the user can do this action.

A new DB method getBoardWriteAccess was added to check user's priority
against the board's heritage and the getPostsBoardInBoardMapping method
was updated to handle a board depth lower than 2 boards.

The byThread route was also updated to return a new property,
'writeAccess', that returns whether the user's priority allows the user
to post to this board based on the board's postable_by heritage.