Full support for newbie and patroller roles
Adds support for the newbie and patroller roles. It also adds a new
priority option for the delete and update permission. The priority
option allows one role to execute the given action on another role if
the priority matches. Currently this is hard coded from patroller to
newbie roles. The two main places this occurs are in the permissions
themselves and in the auth methods that need to handle those
permissions.

Other changes include: returning the post user's priority in byThreads,
and a new pre method that automatically turns all inlined images by a
newbie user into links.