Post Locking
Support for the ability to lock a post so that the original poster
cannot edit or (un)delete the post again. This is ensure that if a
moderator or admin type role modifies a post, it can no longer be
modified by the original poster.

This required adding a new permission to handle th lock and unlock
routes. A new auth method was also created to handle the lock and unlock
routes. The delete auth method was also modified to allow quick locking
a post at the same time a post is deleted. DB methods or retrieving and
setting the post lock were also created. Finally a common method was
created to handle clearing the locked query in the delete route if the
user does not have lock permissions.