Support for meta tag information for threads
Adds support for injecting thread information into the public/index.html
page via handlebars to populate the meta tag for social media sites.
Currently, the forum name and thread title are injected into the title,
while the first post's body is injected into the description tags. For
those sites that handle twitter's labels, the post count and user that
created the thread is also shown.

This was made possible by adding a new route for robots to consume that
loads only the information mentioned above. For regular users, the same
process is applied but once the site loads, they pull down this
information again.